# [CPU] Optimize the unique operator