### Details:
 - *Optimize  unique using HashMap when sorted=false*
 - *...*

### Tickets:
 - *ticket-id*