@dmitry-gorokhov 2024-06-05
Hi @zhaohb. Thanks for the contribution!
Are you doing this optimization in bounds of some specific activity? Maybe there is a ticket with some context?
May I also ask you to share some benchmarking data which confirms perfomance boost?