@zhaohb 2024-06-05
@dmitry-gorokhov 
sorry,  there is no ticket about that. 
I tested the performance with a 256*1910 input, and the tested model contains only a single unique node.
Here is my test code:
```python
import numpy as np
import argparse
import time
import os
import openvino as ov

def main():
    parser = argparse.ArgumentParser()
    parser.add_argument('-d', '--device', dest='device', default='CPU')
    args = parser.parse_args()
    xml_path = './unique.xml'
    assert(xml_path.endswith('.xml'))
    bin_path = xml_path[:xml_path.rfind('.xml')] + '.bin'

    ie = ov.Core()
    net = ie.read_model(xml_path, bin_path)
    exec_net = ie.compile_model(net, args.device)
    dt_dict = {
        'I32': np.int32,
        'FP32': np.float32
    }

    input_dict = {}
    if os.path.exists('unique_input.npy'):
        input_dict['input_1'] = np.load('unique_input.npy')
    else:
        input_dict['input_1'] = np.random.randint(1000, size=(256, 1910))
        np.save("unique_input.npy", input_dict['input_1'])
    print("input: ", input_dict)

    print('Compute...')
    for i in range(10):
        start = time.time()
        outputs = exec_net(inputs=input_dict)
        print('Spend time: %.6f seconds' % (time.time() - start))

if __name__ == '__main__':
    main()
```
before optimization:
```shell
Spend time: 0.100118 seconds
Spend time: 0.094602 seconds
Spend time: 0.092043 seconds
Spend time: 0.094956 seconds
Spend time: 0.091777 seconds
Spend time: 0.091769 seconds
Spend time: 0.091607 seconds
Spend time: 0.091774 seconds
Spend time: 0.094428 seconds
Spend time: 0.093867 seconds

```
after optimization:
```shell
Spend time: 0.017405 seconds
Spend time: 0.010682 seconds
Spend time: 0.007918 seconds
Spend time: 0.008091 seconds
Spend time: 0.007862 seconds
Spend time: 0.007854 seconds
Spend time: 0.007772 seconds
Spend time: 0.007797 seconds
Spend time: 0.007781 seconds
Spend time: 0.007818 seconds
```
So it seems the improvement is quite significant.
Thank you very much.